Change C-style casts to C++ syntax

+++ main.cpp
@@ -18,7 +18,7 @@ int main() {
     unsigned int deck[deckSize] = {
         1, 1, 1, 1, 2, 2, 2, 2, 3, 3, 3, 3
     };
-    std::srand(unsigned(std::time(NULL)));
+    std::srand(static_cast<int>(std::time(NULL)));
     for (unsigned int played = 0; played < rounds; played++) {
         shuffle_deck(deck, deckSize);
         unsigned int n = 0;
@@ -30,7 +30,7 @@ int main() {
             }
         }
     }
-    double winrate = (double)(rounds - lost) / (double)rounds * 100;
+    double winrate = static_cast<double>(rounds - lost) / rounds * 100;
     printf("Lost %d out of %d rounds.\nWinrate is %.4f%%\n", lost,
         rounds, winrate);
     return 0;
